+++ lib/dry/types/core.rb
@@ -52,6 +52,8 @@ module Dry
     # Register :bool since it's common and not a built-in Ruby type :(
     register("bool", self["true"] | self["false"])
     register("strict.bool", self["strict.true"] | self["strict.false"])
+
+    register("object", Definition[::Object].new(::Object))
   end
 end